pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "contract-reviewer"
version = "0.1.0"
description = "AI-powered legal contract review system supporting cloud and local LLM deployment"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.50",
    "chromadb>=0.5",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "fastapi>=0.110",
    "uvicorn>=0.30",
    "typer>=0.12",
    "jinja2>=3.1",
    "python-docx>=1.1",
    "pdfplumber>=0.11",
    "tiktoken>=0.7",
    "diskcache>=5.6",
    "sse-starlette>=2.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
ocr = ["paddleocr>=2.9", "pdf2image>=1.17"]
reranker = ["sentence-transformers>=3.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "respx>=0.21"]

[project.scripts]
contract-review = "contract_reviewer.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/contract_reviewer"]

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "LOG",  # flake8-logging (catches f-strings in log calls)
]
ignore = [
    "E501",  # line length handled by formatter
]

[tool.ruff.lint.isort]
known-first-party = ["contract_reviewer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
